refactor commnents
